requirements.txt
setuptools
langchain
langchain_openai 
python-dotenv
-e .
